requirements.txt
Django==6.0.2
djangorestframework
djangorestframework-simplejwt
psycopg[binary]
